+++ packages/@codotype/ui/src/components/Input/constants.ts
@@ -1,0 +1,2 @@
+export const inputClassName =
+    "px-2 py-5 h-3 font-light text-gray-900 dark:text-gray-200 bg-white dark:bg-gray-900 border border-gray-400 dark:border-gray-700 rounded-md transition-colors w-full";