+++ packages/@codotype/runtime/CHANGELOG.md
@@ -1,3 +1,13 @@
+### 0.8.0-canary.12
+
+-   Updated `yarn.lock` file
+
+### 0.8.0-canary.11
+
+-   Updates `@codotype/core` to `0.8.0-canary.18`
+
+---
+
 ### 0.8.0-canary.10
 
 -   Updates `@codotype/core` to `0.8.0-canary.17`